IrisDataScienceKit
IrisDataScienceKit is a comprehensive data science project that focuses on building, training, and optimizing a machine learning model using the Iris dataset. The main goal of the project is to classify different species of the Iris flower based on their morphological features such as sepal length, sepal width, petal length, and petal width.